+++ SLK_img2pixel/src/process.h
@@ -37,5 +37,6 @@ extern int alpha_threshold;
 
 void process_image(const SLK_RGB_sprite *in, SLK_RGB_sprite *out, SLK_Palette *palette, int sample_mode, int process_mode);
 void sharpen_image(SLK_RGB_sprite *in, SLK_RGB_sprite *out);
+void lowpass_image(SLK_RGB_sprite *in, SLK_RGB_sprite *out);
 
 #endif