+++ SLK_img2pixel/src/utility_win.c
@@ -231,6 +231,7 @@ void dir_output_select(int dither_mode, int sample_mode, int scale_mode, int wid
             char tmp[1028];
             sprintf(tmp,"%s/%s",buffer,file.name);
             SLK_RGB_sprite *in = image_load(tmp);
+            lowpass_image(in,in);
             sharpen_image(in,in);
             if(in!=NULL)
             {