+++ SLK_img2pixel/changelog.md
@@ -10,6 +10,9 @@ Version:                SLK_img2pixel 1.0 - initial release (tba)
 CHANGES:
 
 * fix sharpness not changing at value of 0
+* renamed supersampling to box sampling
+* added lowpass filtering
+* added lanczos sample mode
 
 ----------------------------------------
 Version:                SLK_img2pixel 0.3 - supersampling is amazing (24.01.2021)